+++ validator-string/src/main/java/name/valery1707/jcommander/validators/string/NotEmpty.java
@@ -1,0 +1,15 @@
+package name.valery1707.jcommander.validators.string;
+
+import name.valery1707.jcommander.validators.ValueChecker;
+
+public class NotEmpty extends ValueChecker<CharSequence> {
+	@Override
+	protected boolean check(CharSequence value) {
+		return value.length() > 0;
+	}
+
+	@Override
+	protected String error() {
+		return "be not empty string";
+	}
+}